requirements.txt
flask
torchvision
flask-cors
Pillow